+++ src/com/topodroid/DistoX/ProjectionDialog.java
@@ -198,10 +198,10 @@ class ProjectionDialog extends MyDialog
       NumStation st1 = sh.from;
       NumStation st2 = sh.to;
       if ( st1.show() && st2.show() ) {
-	float x1 = st1.e * cosp + st1.s * sinp; // - dx;
-	float x2 = st2.e * cosp + st2.s * sinp; // - dx;
-	float y1 = st1.v; // - dy;
-	float y2 = st2.v; // - dy;
+	double x1 = st1.e * cosp + st1.s * sinp; // - dx;
+	double x2 = st2.e * cosp + st2.s * sinp; // - dx;
+	double y1 = st1.v; // - dy;
+	double y2 = st2.v; // - dy;
         h1 = DrawingUtil.toSceneX( x1, y1 );
         h2 = DrawingUtil.toSceneX( x2, y2 );
         v1 = DrawingUtil.toSceneY( x1, y1 );
@@ -212,10 +212,10 @@ class ProjectionDialog extends MyDialog
     for ( NumSplay sp : splays ) {
       NumStation st = sp.from;
       if ( st.show() ) {
-	float x1 = st.e * cosp + st.s * sinp; // - dx;
-	float x2 = sp.e * cosp + sp.s * sinp; // - dx;
-	float y1 = st.v; // - dy;
-	float y2 = sp.v; // - dy;
+	double x1 = st.e * cosp + st.s * sinp; // - dx;
+	double x2 = sp.e * cosp + sp.s * sinp; // - dx;
+	double y1 = st.v; // - dy;
+	double y2 = sp.v; // - dy;
         h1 = DrawingUtil.toSceneX( x1, y1 );
         h2 = DrawingUtil.toSceneX( x2, y2 );
         v1 = DrawingUtil.toSceneY( x1, y1 );
@@ -225,8 +225,8 @@ class ProjectionDialog extends MyDialog
     }
     for ( NumStation st : stations ) {
       if ( st.show() ) {
-	float x1 = st.e * cosp + st.s * sinp; // - dx;
-	float y1 = st.v; // - dy;
+	double x1 = st.e * cosp + st.s * sinp; // - dx;
+	double y1 = st.v; // - dy;
         h1 = DrawingUtil.toSceneX( x1, y1 );
         v1 = DrawingUtil.toSceneY( x1, y1 );
         mProjectionSurface.addDrawingStationName( st, h1, v1 );
